requirements.txt
# Core Framework
streamlit>=1.28.0

# PowerPoint Processing
python-pptx>=0.6.21
Pillow>=10.0.0

# AI/LLM
anthropic>=0.18.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# Vector Database
chromadb>=0.4.22

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
